core: implement delete on ttl is zero or above and improve persistence to store ttl

+++ src/core/Cache.h
@@ -12,7 +12,7 @@ public:
   void set(std::string key, std::string value, unsigned int ttl_sec);
   std::string get(std::string key);
   void remove(std::string key);
-  unsigned int ttl(std::string key);
+  unsigned long int ttl(std::string key);
   void save_on_disk();
   void read_from_disk();
